+++ lib/auth/domain/constants.dart
@@ -1,0 +1,3 @@
+class AuthConstants {
+  static const String authSessionKey = "auth_session";
+}
