+++ pubspec.yaml
@@ -31,11 +31,11 @@ dependencies:
   flutter:
     sdk: flutter
 
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   dio: ^5.3.3
+  go_router: ^11.1.2
 
 dev_dependencies:
   flutter_test:
@@ -60,9 +60,9 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-#   assets:
-#     - images/a_dot_burr.jpeg
-#     - images/a_dot_ham.jpeg
+  #   assets:
+  #     - images/a_dot_burr.jpeg
+  #     - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
